+++ src/utils/index.ts
@@ -1,7 +1,6 @@
 export * as Arr from "./arr";
 export * as Dom from "./dom";
 export * as Enum from "./enum";
-export * as Map from "./map";
 export * as Math from "./math";
 export * as Obj from "./obj";
 export * as Str from "./str";