+++ src/utils/map/index.ts
@@ -1,4 +1,0 @@
-
-export function getMapKeys<K, V>(map: Map<K, V>): K[] {
-    return [...map.keys()];
-}