+++ src/utils/obj/index.ts
@@ -3,11 +3,13 @@ import { isObject } from "../../guard";
 export { isObject }
 
 /**
- * <pre>
- * Usage:
- *   hasProperties(obj, ["a", "b"]);          // Gives type Record<string, unknown>
- *   hasProperties(obj, ["a", "b"] as const); // Gives type Record<"a" | "b", unknown>
- * </pre>
+ * ```ts
+ * // Gives type Record<string, unknown>
+ * hasProperties(obj, ["a", "b"]);
+ * 
+ * // Gives type Record<"a" | "b", unknown>
+ * hasProperties(obj, ["a", "b"] as const);
+ * ```
  */
 export function hasProperties<T extends readonly string[]>(obj: unknown, props: T | string[]): obj is Record<T[number], unknown> {
     return isObject(obj) && props.every(p => p in obj);